requirements.txt
shap==0.39.0
streamlit==0.75.0
pandas==1.2.3
numpy==1.19.2
joblib==0.16.0
matplotlib==3.1.3
Pillow==8.2.0
scikit_learn==0.24.1
